Only change TLS settings if required
During module initialisation, calls are made to Get-SXApi to see if
it works without manually enabling TLS v1.2. If it does work - fine,
carry on as normal. If it doesn't work, and TLS v1.2 needs to be
manually set, then this is done only at the start of each API call, then
is reset again immediately afterwards, thus not making any lasting
changes to the PowerShell session.

To facilitate this, the calls to the API have been extracted to a new
private function `Get-SXData`.